Updated the HomeBloc with add and remove locations, and streaming of the locations for the UI by combining from Firebase and Google Places
Implemented the Auth class for use when authenticating users for Add and remove permissions

Implemented a MarkerProvider that creates the marker bitmap for the map. By combining this with the firestore tennislocations and the places search results, this should now only load the map once - when all the data has returned.