Fixed an issue with the add new form not showing when the user selected the text field.
updated the firestore db method to properly add the tennis location including geoflutterfire data.

Made the id of the tennis club optional in the constructor

Created the add and remove methods for tennis clubs and locations in the blocs.

Updated the apiPath class as firestore was seemingly adding the trailing / to the collection names